Bytecode is 16-bits. AMAZING MEMORY USE REDUCTION IN SANITY.
uintptr_t is 64 bits on my system. I assumed that the memory reduction would
be significant, because bytecode takes a lot of memory. But I thought that
signatures and values took up a significant portion of memory. I was wrong.

(Numbers reported by valgrind)
sanity.ly, before: 402,872 (402 KB).

After...

208,370 (208 KB).

It's...very close to half of the entire memory use. This is far above and
beyond what I expected in memory savings. I made sure that this did not break
mod_lily.

Wow. Just...wow.

+++ src/lily_ast.h
@@ -19,7 +19,7 @@ typedef struct {
 
 typedef struct lily_ast_t {
     lily_tree_type tree_type;
-    int line_num;
+    uint16_t line_num;
 
     lily_sig *sig;
     /* This is where the result of evaluating the tree goes. This is used
@@ -88,7 +88,7 @@ typedef struct {
     int save_depth;
 
     lily_raiser *raiser;
-    int *lex_linenum;
+    uint16_t *lex_linenum;
 } lily_ast_pool;
 
 void lily_ast_collect_arg(lily_ast_pool *);
